Authentication working. Added Glucose / Sensor Rx / Tx. Write to control.

+++ app/src/main/java/com/gintechsystems/gincose/messages/AuthChallengeRxMessage.java
@@ -6,7 +6,7 @@ import java.util.Arrays;
  * Created by joeginley on 3/16/16.
  */
 public class AuthChallengeRxMessage extends TransmitterMessage {
-    int opcode = 0x3;
+    byte opcode = 0x3;
     public byte[] tokenHash;
     public byte[] challenge;
